chore: merge vuetify modules and layers
