+++ package.json
@@ -3,7 +3,7 @@
   "type": "module",
   "version": "0.4.3",
   "packageManager": "pnpm@8.6.7",
-  "description": "Zero-Config Nuxt Module for Vuetify ",
+  "description": "Zero-Config Nuxt Module for Vuetify",
   "author": "userquin <userquin@gmail.com>",
   "license": "MIT",
   "homepage": "https://github.com/userquin/vuetify-nuxt-module#readme",
